refactor(users): extract content and detail logic from UserService into separate services
- Moved content-related logic to ContentService
- Moved user profile operations to UserDetailsService
- Introduced shared ApiResponse<T> for consistent API response formatting
- Improves separation of concerns and maintainability
