+++ src/common/interfaces/api-response.interface.ts
@@ -1,0 +1,7 @@
+import { PaginationMeta } from "./pagination-meta.interface";
+
+export interface ApiResponse<T = any> {
+  message: string;
+  data: T;
+  meta?: PaginationMeta;
+}
