+++ src/common/interfaces/pagination-meta.interface.ts
@@ -1,0 +1,9 @@
+export interface PaginationMeta {
+  total: number;
+  page: number;
+  limit: number;
+  totalPages: number;
+  query?: string;
+  search?: string;
+  category_id?: number;
+}
