+++ src/users/__http_tests__/users.threads.http
@@ -1,7 +1,7 @@
 @port = {{$dotenv API_PORT}}
 @baseUrl = http://localhost:{{port}}
-@accessToken = eyJhbGciOiJQUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjEsInRva2VuVHlwZSI6ImFjY2VzcyIsInJvbGUiOiJTVVBFUl9BRE1JTiIsImlhdCI6MTc0NTAyMzgyMSwiZXhwIjoxNzQ1MDI0NzIxfQ.Zuxau6k1xYIo-Tu0xTAUXnwFV6KgbM_rTElqpWM_IVdfM4o2rVi4LDHCx3NXKUcKg711xt9xkZSfnFnEL_-KClxPxatvEVhDJpl6U-7DWUhfk5DhcDMUutFYjf6xROGMEs8ELNo2c4Rrw3dL1rwjgItKg6iJu8DAdbMW4iGKcK7hciF-1BiQ77LIMHe8kvX2lV7wWLT_wTI13hmUDjhFxTzmBKImR1JoRwSX_d_nqnnBe0nZ5iv7t_Ivv7FrRLIs0cPXgZuo6q03Wif4pNy5Er0hW0CigBXp12dIaPBCxSPYYlVFzQqmLWJ4IbIbiv62dCeauDfKORDgPFAbB1sstD2UjcZV68QxnIA4103_-WmuDUmPjvCbqGiHw9QASY84jWp8xUEd7AXhDQgWYomXjRnoigXKNRn40CVqZV-rgIaiFuD64VuTP9GN0s8TOHZdLL7da_XHix7-CP8FVk3bu1S0LLm1158IZTf5S5Q78x696CcwESYT3_pZLdCgZ9WmbH4jwPrH4AiMeoO55zAZFcaVe03f7n6jvYzhEURL6QFT_KEfLpM39V6p7khEgjcQhTz5uFl72mcdBVpBRPFQ01TYbF1PTBm3cYv0bbRRimeqMvt0Eh7NUVBmwO-NUQqekNfXESxg4EGoFsF3sVmxZgDrtaMlWfg5MZmC6vLAc_E
+@accessToken = eyJhbGciOiJQUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjEsInRva2VuVHlwZSI6ImFjY2VzcyIsInJvbGUiOiJTVVBFUl9BRE1JTiIsImlhdCI6MTc0NjM3NDA1MSwiZXhwIjoxNzQ2Mzc0OTUxfQ.p6R9aHMyqhBIT5HZ53j2SBcToPLLW-KH1HI3_PSRYiRc1KqzCbxkK4Nv97KcfZorDFRICAJywE6O-KXVK_r9L5ApYEcbvcYTkU3m0AXS7c3H8LHGXd3e2LMuH2TkJvC3F7tCejF-nELR0xntNvMLkqT0BiZvkrXJ3Jlrqz9zMVp4x9Kn9aH5kqjq7hZhTNAsPKTuiuAK-0FAWkKCigviZ0Z5dR-VLS28Qwi162mndHU74ARwONooNfnltfeLVXiuu3EW3_WmfTpXmM1tj3wx3X-EkoyS-2qUBab2PTSAK_V0dVUsxCuBoT2Pjfh6LXTN5jO_FiDzkU2ZbnXR1ViJ7gSPN1TGsgRjC67aX4Vx4fFu_jLwf-uAW3XDo9ylRMuRy2KCpjv89pelgqlZk7Ssm_4uspwQ7BFSaECKTNI8P0oNZ_k4ayBGeNOW5BPIAGmrqPw-HgUN4qkN3yTqs5QHCQd83mFfr1nhSG0ILNPjOCtgPdd8LmN_LRsyNuJ4cuzw3OIYWkAZTMqPuUDLwl7ue6o5U7srAkwEbVJ00_YwVGwv0UOEmpzLzpzLlTILilbITuLT63zm4vNrLOyfALf3j_sh79MCi6I3cvVKWvl0MHC8KD6sDaJiTo2FvuOpYxGdixz6bD3Sq3geZIXt9qgA_6bw2mIY4bwdvrw7rPkwVnw
 
 ### Get user threads
-GET {{baseUrl}}/users/me/threads?search=Title&category_id=2
+GET {{baseUrl}}/users/me/threads
 Authorization: Bearer {{accessToken}}
